Rename utils to machine and update some docs
