+++ models/rasmachine/run_machine.sh
@@ -13,4 +13,4 @@ cd "$DIR"
 export PYTHONPATH=$PYTHONPATH:../../../indra
 echo $PYTHONPATH
 log_file="$model_name/log.txt"
-indra machine run_with_search $model_name >> $log_file 2>> $log_file
+python -m indra.tools.machine run_with_search $model_name >> $log_file 2>> $log_file
